Simulated Epidemiological Dataset for Togo

A dataset representing simulated epidemiological data for malaria and cholera across various states in Togo from 2018-2022.

Usage

data("fake_epi_df_togo")

Format

A data frame with 900 rows and 8 variables:

country

The country of the data. Always 'Togo' in this dataset.

district

The state in Togo where the data was collected. Includes 15 different districts

region

The region in Togo where the data was collected. Includes 4 different regions

month

The month when the data was recorded, ranging from 1 to 12.

year

The year when the data was recorded, ranging from 2018 to 2022.

malaria_tests

The number of tests conducted for malaria. Includes random missing values and outliers.

malaria_cases

The number of confirmed cases for malaria. Includes random missing values and outliers.

cholera_tests

The number of tests conducted for cholera. Includes random missing values and outliers.

cholera_cases

The number of confirmed cases for cholera. Includes random missing values and outliers.

Details

The dataset was generated using various steps:

The dataset was initiated using `dplyr::expand_grid()` to create a combination of districts, region, months, and years.

Random values were generated for malaria_tests, malaria_cases, cholera_tests, and cholera_cases using a normal distribution.

Outliers were introduced for 5 malaria_cases, and cholera_cases.

24 to NA.

The number of 'cases' was adjusted to make sure they are less than the number of 'tests'.

Any negative 'cases' values were converted to their absolute values.

Examples

data('fake_epi_df_togo')

head(fake_epi_df_togo)
#> # A tibble: 6 × 9
#>   country region  district  year month malaria_tests malaria_cases cholera_tests
#>   <chr>   <chr>   <chr>    <int> <int>         <dbl>         <dbl>         <dbl>
#> 1 Togo    Centra… Centrale  2018     1           113            NA            NA
#> 2 Togo    Centra… Centrale  2018     2           129            10            75
#> 3 Togo    Centra… Centrale  2018     3           134            21            NA
#> 4 Togo    Centra… Centrale  2018     4            84             7            85
#> 5 Togo    Centra… Centrale  2018     5           114            NA           139
#> 6 Togo    Centra… Centrale  2018     6            89             1           111
#> # ℹ 1 more variable: cholera_cases <dbl>